Voice Analyzer Improved and File Observation Enhanced
- Implemented file observation for Eitaa, Downloads, Music, and Recordings directories.
- Added a progress indicator during file upload.
- Enhanced error handling and reporting during file upload and analysis.
- Displayed upload progress percentage with higher precision.
- Updated voice analysis responses to exclude redundant transcriptions.
- Fixed a bug where the transcription was sometimes missing in the response.
- Added a loading indicator during app initialization.
- Replaced the help text on the WeeklyReportPage with a more concise message.

+++ app/src/main/java/ir/saltech/sokhanyar/util/ProgressRequestBody.java
@@ -58,7 +58,7 @@ public class ProgressRequestBody extends RequestBody {
     }
 
     public interface UploadCallbacks {
-        void onProgressUpdate(int percentage);
+        void onProgressUpdate(float percentage);
 
         void onError();
 
@@ -76,7 +76,7 @@ public class ProgressRequestBody extends RequestBody {
 
         @Override
         public void run() {
-            mListener.onProgressUpdate((int) (100 * mUploaded / mTotal));
+            mListener.onProgressUpdate((float) (100 * mUploaded) / mTotal);
         }
     }
 }
